refactor test ingestor
